move customField to components
