+++ modules/Settings/TicketField/Field.js
@@ -8,7 +8,7 @@ import { useMutation, useQuery } from 'react-query'
 import { DocumentTitle } from 'modules/utils/DocumentTitle'
 import { RadioGroup } from 'modules/components/Radio'
 import Divider from 'modules/components/Divider'
-import { fieldType, includeOptionsType } from './CustomField'
+import { fieldType, includeOptionsType } from 'modules/components/CustomField'
 import { useAppContext } from 'modules/context'
 import _ from 'lodash'
 import Preview from './Preview'
