+++ modules/Settings/TicketField/index.module.scss
@@ -81,7 +81,4 @@
 .displayType {
   font-size: 16px;
 }
-.optionItem {
-  min-width: 50%;
-  margin-right: 0 !important;
-}
+
