+++ modules/Settings/TicketForm/FormPage.js
@@ -21,7 +21,7 @@ import { useTranslation } from 'react-i18next'
 import { useAppContext } from 'modules/context'
 import { DocumentTitle } from 'modules/utils/DocumentTitle'
 import NoData from 'modules/components/NoData'
-import { includeOptionsType } from '../TicketField/CustomField'
+import { includeOptionsType } from 'modules/components/CustomField'
 import { useFormId } from './'
 import Preview from './Preview'
 import styles from './index.module.scss'
