+++ modules/components/CustomField/index.module.scss
@@ -1,0 +1,4 @@
+.optionItem {
+  min-width: 50%;
+  margin-right: 0 !important;
+}
